+++ include/learning/forests/obliviousmart.h
@@ -41,7 +41,7 @@ class ObliviousMart : public Mart {
   /// \param minleafsupport Minimum number of instances in each leaf.
   /// \param esr Early stopping if no improvement after \esr iterations
   /// on the validation set.
-  ObliviousMart(size_t ntrees, float shrinkage, size_t nthresholds,
+  ObliviousMart(size_t ntrees, double shrinkage, size_t nthresholds,
                 size_t treedepth, size_t minleafsupport,
                 size_t esr)
       : Mart(ntrees, shrinkage, nthresholds, 1 << treedepth, minleafsupport,
